+++ spec/state_spec.rb
@@ -27,64 +27,64 @@ module Pathway
       end
     end
 
-    describe '#unwrap' do
+    describe '#use' do
       before { state.update(val: 'RESULT', foo: 99, bar: 131) }
       it 'fails if no block is provided' do
-        expect { state.unwrap }.to raise_error('a block must be provided')
+        expect { state.use }.to raise_error('a block must be provided')
       end
 
       context 'when a block is provided' do
         it 'passes specified values using only keyword params', :aggregate_failures do
-          expect(state.unwrap {|val:| val }).to eq('RESULT')
-          expect(state.unwrap {|foo:| foo }).to eq(99)
-          expect(state.unwrap {|val:, bar:| [val, bar] })
+          expect(state.use {|val:| val }).to eq('RESULT')
+          expect(state.use {|foo:| foo }).to eq(99)
+          expect(state.use {|val:, bar:| [val, bar] })
             .to eq(['RESULT', 131])
         end
 
         it 'passes no arguments if no keyword or positional params are defined' do
-          expect(state.unwrap { 77 }).to eq(77)
+          expect(state.use { 77 }).to eq(77)
         end
 
         it 'passes specified values using only positional params', :aggregate_failures do
-          expect(state.unwrap {|val| val }).to eq('RESULT')
-          expect(state.unwrap {|foo| foo }).to eq(99)
-          expect(state.unwrap {|val, bar| [val, bar] })
+          expect(state.use {|val| val }).to eq('RESULT')
+          expect(state.use {|foo| foo }).to eq(99)
+          expect(state.use {|val, bar| [val, bar] })
         end
 
         it 'fails if positional and keyword params are both defined', :aggregate_failures do
-          expect { state.unwrap {|pos, input:| } }
+          expect { state.use {|pos, input:| } }
             .to raise_error('cannot mix positional and keyword arguments')
         end
 
         it 'fails if using rest param', :aggregate_failures do
-          expect { state.unwrap {|*input| } }
+          expect { state.use {|*input| } }
             .to raise_error('rest arguments are not supported')
-          expect { state.unwrap {|input, *args| args } }
+          expect { state.use {|input, *args| args } }
             .to raise_error('rest arguments are not supported')
         end
 
         it 'fails if using keyrest param', :aggregate_failures do
-          expect { state.unwrap {|**kargs| kargs } }
+          expect { state.use {|**kargs| kargs } }
             .to raise_error('rest arguments are not supported')
-          expect { state.unwrap {|input:, **kargs| kargs } }
+          expect { state.use {|input:, **kargs| kargs } }
             .to raise_error('rest arguments are not supported')
         end
 
         context 'that takes a block argument' do
           it 'fails if it has positional and keyword params' do
-            expect { state.unwrap {|input, val:, &bl| } }
+            expect { state.use {|input, val:, &bl| } }
               .to raise_error('cannot mix positional and keyword arguments')
           end
 
           it 'does not fails if only has keyword params', :aggregate_failures do
-            expect(state.unwrap {|val:, &bl| val }).to eq('RESULT')
-            expect(state.unwrap {|val:, &_| val }).to eq('RESULT')
-            expect(state.unwrap {|&_| 77 }).to eq(77)
+            expect(state.use {|val:, &bl| val }).to eq('RESULT')
+            expect(state.use {|val:, &_| val }).to eq('RESULT')
+            expect(state.use {|&_| 77 }).to eq(77)
           end
 
           it 'does not fails if only has positional params', :aggregate_failures do
-            expect(state.unwrap {|val, &bl| val }).to eq('RESULT')
-            expect(state.unwrap {|val, &_| val }).to eq('RESULT')
+            expect(state.use {|val, &bl| val }).to eq('RESULT')
+            expect(state.use {|val, &_| val }).to eq('RESULT')
           end
         end